made explore bar bold on click
